# docs/api-documentation.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1.0/posts/profile/me": {
      "get": {
        "description": "Retrieves a paginated list of posts created by the authenticated user",
        "operationId": "PostController_getProfilePosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of posts per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get user profile posts",
        "tags": [
          "Posts"
        ]
      }
    },
...
    "/api/v1.0/profile/me": {
      "get": {
        "description": "Returns the profile of the currently authenticated user.",
        "operationId": "ProfileController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Profile"
        ]
      },
      "patch": {
        "description": "Updates the profile of the currently authenticated user.",
        "operationId": "ProfileController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Profile"
        ]
      }
    },
    "/api/v1.0/profile/user/{userId}": {
      "get": {
        "description": "Returns the profile of a specific user by their user ID.",
        "operationId": "ProfileController_getProfileByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "The ID of the user",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get user profile by user ID",
        "tags": [
          "Profile"
        ]
      }
    },
    "/api/v1.0/profile/username/{username}": {
      "get": {
        "description": "Returns the profile of a specific user by their username.",
        "operationId": "ProfileController_getProfileByUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "The username of the user",
            "schema": {
              "example": "john_doe",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get user profile by username",
        "tags": [
          "Profile"
        ]
      }
    },
    "/api/v1.0/profile/search": {
      "get": {
        "description": "Search for user profiles by partial match on username or name. Supports pagination.",
        "operationId": "ProfileController_searchProfiles",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Search query to match against username or name",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profiles found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Search profiles by username or name",
        "tags": [
          "Profile"
        ]
      }
    }
  },
  "info": {
    "title": "Hankers",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "error",
              "fail"
            ],
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "error": {
            "type": "object",
            "nullable": true,
            "example": "Bad Request",
            "description": "Optional error details or the type of error"
          }
        },
        "required": [
          "status",
          "message",
          "error"
        ]
      },
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error",
              "fail"
            ],
            "example": "success",
            "description": "The status of the response"
          },
          "message": {
            "type": "string",
            "example": "Operation successful",
            "description": "A descriptive message about the response"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "The data payload of the response"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      },
...
      "UserInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john@example.com"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "example": "USER",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "created_at"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Profile ID",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "description": "User ID associated with this profile",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John Doe"
          },
          "birth_date": {
            "format": "date-time",
            "type": "string",
            "description": "User birth date",
            "example": "1990-01-01T00:00:00.000Z"
          },
          "profile_image_url": {
            "type": "string",
            "description": "Profile image URL",
            "example": "https://example.com/profile.jpg"
          },
          "banner_image_url": {
            "type": "string",
            "description": "Banner image URL",
            "example": "https://example.com/banner.jpg"
          },
          "bio": {
            "type": "string",
            "description": "User bio",
            "example": "Software developer"
          },
          "location": {
            "type": "string",
            "description": "User location",
            "example": "San Francisco, CA"
          },
          "website": {
            "type": "string",
            "description": "User website",
            "example": "https://johndoe.com"
          },
          "is_deactivated": {
            "type": "boolean",
            "description": "Whether the profile is deactivated",
            "example": false
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Profile creation timestamp",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Profile last update timestamp",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "User": {
            "description": "Associated user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfoDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "birth_date",
          "created_at",
          "updated_at",
          "User"
        ]
      },
      "GetProfileResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Profile retrieved successfully"
          },
          "data": {
            "description": "Profile data",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileResponseDto"
              }
            ]
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      },
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total number of results",
            "example": 25
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages",
            "example": 3
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "SearchProfileResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Profiles found successfully"
          },
          "data": {
            "description": "Array of matching profiles",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileResponseDto"
            }
          },
          "metadata": {
            "description": "Pagination metadata",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetadata"
              }
            ]
          }
        },
        "required": [
          "status",
          "message",
          "data",
          "metadata"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe",
            "maxLength": 30
          },
          "birth_date": {
            "type": "string",
            "description": "The birth date of the user",
            "example": "1990-01-01",
            "format": "date"
          },
          "profile_image_url": {
            "type": "string",
            "description": "URL of the user profile image",
            "example": "https://example.com/profile.jpg",
            "maxLength": 255
          },
          "banner_image_url": {
            "type": "string",
            "description": "URL of the user banner image",
            "example": "https://example.com/banner.jpg",
            "maxLength": 255
          },
          "bio": {
            "type": "string",
            "description": "User biography",
            "example": "Software developer passionate about clean code",
            "maxLength": 160
          },
          "location": {
            "type": "string",
            "description": "User location",
            "example": "San Francisco, CA",
            "maxLength": 100
          },
          "website": {
            "type": "string",
            "description": "User website URL",
            "example": "https://johndoe.com",
            "maxLength": 100
          }
        }
      },
      "UpdateProfileResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response status",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Profile updated successfully"
          },
          "data": {
            "description": "Updated profile data",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileResponseDto"
              }
            ]
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      }
    }
  }
}