# src/profile/dto/update-profile-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { ProfileResponseDto } from './profile-response.dto';

export class UpdateProfileResponseDto {
  @ApiProperty({
    description: 'Response status',
    example: 'success',
  })
  status: string;

  @ApiProperty({
    description: 'Response message',
    example: 'Profile updated successfully',
  })
  message: string;

  @ApiProperty({
    description: 'Updated profile data',
    type: ProfileResponseDto,
  })
  data: ProfileResponseDto;
}